main class added

+++ src/ir/ac/aut/ceit/paint/Circle.java
@@ -10,7 +10,7 @@ import javax.swing.*;
 public class Circle implements Shape {
     private double radius;
 
-    public Circle(int radius) {
+    public Circle(double radius) {
         this.radius = radius;
     }
